A bunch of refactoring; add a loading spinner
