+++ app/js/components/LabOrdersList/Cell.jsx
@@ -112,4 +112,4 @@ function Cell({
   }
 }
 
-export default Cell;
+export default React.memo(Cell);
