+++ app/js/components/LabResultsList/index.js
@@ -1,0 +1,3 @@
+import LabResultsList from "./LabResultsList";
+
+export default LabResultsList;
